+++ src/characterSelect.c
@@ -33,8 +33,8 @@ void character_Select_Init()
 	CP_TEXT_ALIGN_VERTICAL vertical = CP_TEXT_ALIGN_V_MIDDLE;
 	CP_Settings_TextAlignment(horizontal, vertical);
 	CP_Settings_TextSize(35.0f);
-	zWidth = CP_System_GetWindowWidth();
-	zHeight = CP_System_GetWindowHeight();
+	zWidth = (float) CP_System_GetWindowWidth();
+	zHeight = (float)CP_System_GetWindowHeight();
 }
 
 
@@ -54,21 +54,21 @@ void character_Select_Update()
 		// Set up the character select screen 
 		CP_Settings_Fill(CP_Color_Create(0, 0, 0, 255));
 		CP_Font_DrawText("Choose your character", zWidth / 2.0f, zHeight / 2.0f - 200);
-		CP_Image_Draw(gunPlayer, zWidth / 2.0f - 100, zHeight / 2.0f, CP_Image_GetWidth(gunPlayer), CP_Image_GetHeight(gunPlayer), 255);
-		CP_Image_Draw(swordPlayer, zWidth / 2.0f + 100, zHeight / 2.0f, CP_Image_GetWidth(swordPlayer), CP_Image_GetHeight(swordPlayer), 255);
+		CP_Image_Draw(gunPlayer, zWidth / 2.0f - 100, zHeight / 2.0f, (float)CP_Image_GetWidth(gunPlayer), (float) CP_Image_GetHeight(gunPlayer), 255);
+		CP_Image_Draw(swordPlayer, zWidth / 2.0f + 100, zHeight / 2.0f, (float)CP_Image_GetWidth(swordPlayer), (float) CP_Image_GetHeight(swordPlayer), 255);
 	
 		//Pick the character, then proceed to Howtoplay page.
 		if (CP_Input_MouseClicked()) {
 			CP_Vector mouseClickPos = CP_Vector_Set(CP_Input_GetMouseX(), CP_Input_GetMouseY());
 			
-				if (IsAreaClicked(zWidth / 2.0f - 100, zHeight / 2.0f, CP_Image_GetWidth(gunPlayer) + 10, CP_Image_GetHeight(gunPlayer) + 10 , mouseClickPos.x, mouseClickPos.y) == 1)
+				if (IsAreaClicked(zWidth / 2.0f - 100, zHeight / 2.0f, (float) CP_Image_GetWidth(gunPlayer) + 10, (float) CP_Image_GetHeight(gunPlayer) + 10 , mouseClickPos.x, mouseClickPos.y) == 1)
 				{
 					playerNum = 1;
 					CP_Engine_SetNextGameState(how_To_play_Init, how_To_play_Update, how_To_play_Exit);
 					/*panelDisplay = 0;*/
 				}
 
-				if (IsAreaClicked(zWidth / 2.0f + 100, zHeight / 2.0f, CP_Image_GetWidth(swordPlayer) +10, CP_Image_GetHeight(swordPlayer) + 10 , mouseClickPos.x, mouseClickPos.y) == 1)
+				if (IsAreaClicked(zWidth / 2.0f + 100, zHeight / 2.0f, (float)CP_Image_GetWidth(swordPlayer) +10, (float) CP_Image_GetHeight(swordPlayer) + 10 , mouseClickPos.x, mouseClickPos.y) == 1)
 				{
 					playerNum = 2;
 					CP_Engine_SetNextGameState(how_To_play_Init, how_To_play_Update, how_To_play_Exit);
